requirements.txt
dlib==19.22.1
keras==2.8.0
librosa==0.9.1
matplotlib==3.5.1
numpy~=1.21.4
opencv_python==4.5.5.62
Pillow~=8.4.0
tensorflow==2.7.0

scikit-learn~=1.0.2
opencv-python~=4.5.5.62